+++ package.json
@@ -16,7 +16,7 @@
     "conform": "conform",
     "fix-dependencies": "syncpack fix-mismatches",
     "format-package": "syncpack format",
-    "postinstall": "node -r @swc-node/register tools/executors/toolbox/src/toolbox.ts",
+    "postinstall": "tsx tools/executors/toolbox/src/main.ts",
     "install-playwright": "playwright install",
     "lint": "nx run-many --target=lint --fix",
     "lint:changed": "nx run-many --target=lint --projects=$(pnpm run -w --silent changed-packages) --fix",
@@ -25,7 +25,7 @@
     "prepare": "[ \"$HUSKY\" != \"0\" ] && husky install || true",
     "serve": "serve",
     "test": "nx run-many --target=test",
-    "toolbox": "node -r @swc-node/register tools/executors/toolbox/src/toolbox.ts",
+    "toolbox": "tsx tools/executors/toolbox/src/main.ts",
     "vitest": "nx run-many --target=vitest",
     "watch": "./node_modules/.bin/nx watch --all -- pnpm -w nx compile \"\\$NX_PROJECT_NAME\"",
     "watch-build": "./node_modules/.bin/nx watch --all -- pnpm -w nx build \"\\$NX_PROJECT_NAME\""
