+++ packages/core/protocols/project.json
@@ -42,11 +42,11 @@
       "executor": "@nx/js:tsc",
       "options": {
         "main": "{projectRoot}/src/index.ts",
-        "outputPath": "packages/core/protocols/dist/esm",
+        "outputPath": "{projectRoot}/dist/esm",
         "transformers": [
           "@dxos/log-hook/transformer"
         ],
-        "tsConfig": "packages/core/protocols/tsconfig.json"
+        "tsConfig": "{projectRoot}/tsconfig.json"
       },
       "outputs": [
         "{options.outputPath}"
