+++ tools/executors/toolbox/tsconfig.json
@@ -2,6 +2,8 @@
   "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "emitDeclarationOnly": false,
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
     "outDir": "dist"
   },
   "include": [
